refactores the shit out of this app
